+++ RELEASE-NOTES.md
@@ -12,6 +12,7 @@ as release notes for end user on a Recalbox upgrade.
 - Add the possibility to sort games by release date.
 - RRGBD:
   - Add new theme for recalbox rgb dual with 240p optimisations
+  - Add HDMI priority option in recalbox rgb dual to switch between Scart and HDMI
   - Add neogeocd, picodrive and pcsx rearmed, cores for rrgbd
   - Add snes9x2002, 2005, 2010 and bsnes support for rrgbd
   - Add region selection for satellaview, sufami, amiga 600, amiga 1200
@@ -46,6 +47,7 @@ as release notes for end user on a Recalbox upgrade.
 - Fix bad overlay showing when save on exit is enabled on retroarch
 - Fix support for Ultimarc IPAC 2
 - Fix internal scraper saving path before actual media
+- Fix resolution selection in frontend for Raspberry Pi 4/400/3/Zero2 (KMS)
 
 ## Version 8.1.1-Electron